fix(slack): fix channel trigger
