+++ packages/pieces/community/slack/src/lib/common/utils.ts
@@ -75,3 +75,42 @@ export function getFirstFiveOrAll(array: unknown[]) {
     return array.slice(0, 5);
   }
 }
+
+/**
+* Parse a message text to extract command and arguments
+*/
+export function parseCommand(
+ text: string,
+ botUserId: string,
+ validCommands: string[]
+): { command: string; args: string[] } | null {
+ if (!botUserId) {
+   return null;
+ }
+
+ // Check if the message mentions the bot
+ const mentionRegex = new RegExp(`<@${botUserId}>\\s+(.+)`, 's');
+ const mentionMatch = text.match(mentionRegex);
+
+ if (!mentionMatch) {
+   return null;
+ }
+
+ // Extract the text after the mention
+ const commandText = mentionMatch[1].trim();
+
+ // Split into command and arguments (first word is command, rest are args)
+ const parts = commandText.split(/\s+/);
+ const command = parts[0].toLowerCase();
+ const args = parts.slice(1);
+
+ // Check if it's a valid command
+ if (!validCommands.includes(command)) {
+   return null;
+ }
+
+ return {
+   command,
+   args,
+ };
+}
